Cloud-Computing-Personality-Website/backend/requirements.txt
pandas==1.1.5
openpyxl==3.0.7
xlrd==2.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
numpy==1.19.5
torch==1.9.0
transformers==4.10.3
sentencepiece==0.1.95
protobuf==3.17.3
psycopg2==2.9.1
python-dotenv==0.17.1
gunicorn==20.1.0
gevent==21.8.0
greenlet==1.1.2
locust==2.8.6
Flask-Migrate==3.1.0
